pyproject.toml
[project]
name = "insurance-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.116.1",
    "fitz>=0.0.1.dev2",
    "openai>=1.107.3",
    "pydantic-settings>=2.10.1",
    "pymupdf>=1.26.4",
    "pypdf2>=3.0.1",
    "python-multipart>=0.0.20",
    "uvicorn>=0.35.0",
]
